+++ backend/src/main/java/com/popflix/auth/AuthenticationController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.popflix.controller.UserAlreadyExistsException;
+import com.popflix.repository.UserAlreadyExistsException;
 
 @RestController
 @RequestMapping("/api/v1/auth")
