+++ backend/src/main/java/com/popflix/auth/AuthenticationService.java
@@ -7,9 +7,9 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.popflix.config.JwtService;
-import com.popflix.controller.UserAlreadyExistsException;
 import com.popflix.model.Role;
 import com.popflix.model.User;
+import com.popflix.repository.UserAlreadyExistsException;
 import com.popflix.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
